+++ modules/fabric_1.20.1/src/main/java/com/armilp/ifreq/common/registry/ModItems.java
@@ -1,6 +1,7 @@
 package com.armilp.ifreq.common.registry;
 
 import com.armilp.ifreq.MainEZ;
+import com.armilp.ifreq.common.config.IfreqConfig;
 import com.armilp.ifreq.common.items.ItemWalkieTalkie;
 import net.minecraft.item.Item;
 import net.minecraft.registry.Registries;
@@ -10,8 +11,12 @@ import net.minecraft.util.Identifier;
 public class ModItems {
 
     public static final Item WALKIE_TALKIE = new ItemWalkieTalkie(
-            new Item.Settings(), 1000.0
+            new Item.Settings(), "walkie_talkie"
     );
+
+    static {
+        IfreqConfig.preRegister("walkie_talkie", 1000.0);
+    }
 
     public static void register() {
         Registry.register(