+++ modules/fabric_1.20.1/changelog.md
@@ -3,5 +3,6 @@
 ### Added
   - 3d model for the walkie-talkie item
   - walkie-talkie gui can be open while in the blet slot of the Accessories mod hitting R key
+  - Config for editing range of the walkie-talkie
 
 This is a beta version, so expect some bugs and missing features.